Commons Update 0.0.27

+++ src/test/java/de/aservo/confapi/jira/service/MailServerServiceTest.java
@@ -46,7 +46,7 @@ public class MailServerServiceTest {
         assertEquals(smtpMailServer.getName(), bean.getName());
         assertEquals(smtpMailServer.getDescription(), bean.getDescription());
         assertEquals(smtpMailServer.getHostname(), bean.getHost());
-        assertEquals(smtpMailServer.getTimeout(), bean.getTimeout());
+        assertEquals(smtpMailServer.getTimeout(), (long) bean.getTimeout());
         assertEquals(smtpMailServer.getUsername(), bean.getUsername());
         assertNull(bean.getPassword());
         assertEquals(smtpMailServer.getDefaultFrom(), bean.getFrom());
@@ -133,7 +133,7 @@ public class MailServerServiceTest {
         assertEquals(popMailServer.getName(), bean.getName());
         assertEquals(popMailServer.getDescription(), bean.getDescription());
         assertEquals(popMailServer.getHostname(), bean.getHost());
-        assertEquals(popMailServer.getTimeout(), bean.getTimeout());
+        assertEquals(popMailServer.getTimeout(), (long) bean.getTimeout());
         assertEquals(popMailServer.getUsername(), bean.getUsername());
         assertNull(bean.getPassword());
         assertEquals(popMailServer.getMailProtocol().getProtocol(), bean.getProtocol());
